adding better error handling to thought controllers

+++ controllers/thought-controller.js
@@ -4,7 +4,7 @@ const thoughtController = {
   async getAllThoughts(req, res) {
     try {
       const thoughtData = await Thought.find({});
-      if (!thoughtData) {
+      if (thoughtData.length < 1) {
         console.log("You need to create a thought first");
         res.status(500).json({ message: "You need to create a thought first" });
         return;
@@ -18,8 +18,9 @@ const thoughtController = {
 
   async getThought(req, res) {
     try {
-      const thoughtData = await Thought.findOne({ _id: req.params.id })
-      .populate({
+      const thoughtData = await Thought.findOne({
+        _id: req.params.id,
+      }).populate({
         path: "user",
       });
       if (!thoughtData) {
@@ -36,6 +37,7 @@ const thoughtController = {
 
   async updateThought(req, res) {
     try {
+      const thought = await Thought.findOne({ _id: req.params.id });
       const thoughtData = await Thought.findOne({
         _id: req.params.id,
       });
@@ -45,6 +47,11 @@ const thoughtController = {
       }
       if (req.body.thoughtText) {
         thoughtData.thoughtText = req.body.thoughtText;
+      }
+      if (!thought) {
+        console.log("No thought found with that id");
+        res.status(500).json({ message: "No thought found with that id" });
+        return;
       }
       thoughtData.userId = req.body.userId;
       thoughtData.save();
@@ -58,15 +65,15 @@ const thoughtController = {
   async newThought(req, res) {
     try {
       const thoughtData = await Thought.create(req.body);
+      const user = await User.findOne({ _id: req.body.userId });
+      if (!user) {
+        console.log("No user found with that id");
+        return res.status(404).json({ message: "No user found with that id" });
+      }
       const userData = await User.updateOne(
         { _id: req.body.userId },
-        { $push: { thoughts: thoughtData } },
-        { new: true }
+        { $push: { thoughts: thoughtData } }
       );
-      if (!userData) {
-        console.log("No user found with that id");
-        res.status(404).json({ message: "No user found with that id" });
-      }
       res.json(userData);
     } catch (error) {
       console.log(error);
@@ -75,29 +82,26 @@ const thoughtController = {
   },
 
   async deleteThought(req, res) {
-    const thoughtData = await Thought.deleteOne({ _id: req.params.id });
-    if (!thoughtData) {
+    const thought = await Thought.findOne({ _id: req.params.id });
+    if (!thought) {
       console.log("No thought by that id");
       res.status(500).json({ message: "No thought by that id" });
       return;
     }
-    res.json(thoughtData);
+    const thoughtData = await Thought.deleteOne({ _id: req.params.id });
+    res.json(thought);
   },
 
   async deleteAllThoughts(req, res) {
     try {
-      const thoughtData = await Thought.deleteMany({});
-      if (!thoughtData) {
-        console.log(
-          "unable to delete records at this time due to internal server error"
-        );
-        return res
-          .status(500)
-          .json({
-            message:
-              "unable to delete records at this time due to internal server error",
-          });
+      const thought = await Thought.find();
+      if (thought.length < 1) {
+        console.log("No thoughts exist in the database");
+        return res.status(500).json({
+          message: "No thoughts exist in the database",
+        });
       }
+      const thoughtData = await Thought.deleteMany({});
       res.json(thoughtData);
     } catch (error) {
       console.log(error);
@@ -107,15 +111,16 @@ const thoughtController = {
 
   async addReaction(req, res) {
     try {
+      const thought = await Thought.findOne({ _id: req.params.id });
+      if (!thought) {
+        console.log("No thought by that id");
+        return res.status(500).json({ message: "No thought by that id" });
+      }
       const thoughtData = await Thought.updateOne(
         { _id: req.params.id },
         { $push: { reactions: req.body } }
       );
-      if (!thoughtData) {
-        console.log("No thought by that id");
-        return res.status(500).json({ message: "No thought by that id" });
-      }
-      res.json(thoughtData);
+      res.json(thought);
     } catch (error) {
       console.log(error);
       res.status(500).json(error);
@@ -124,15 +129,16 @@ const thoughtController = {
 
   async deleteReaction(req, res) {
     try {
+      const thought = await Thought.findOne({ _id: req.params.id });
+      if (!thought) {
+        console.log("No thought by that id");
+        return res.status(500).json({ message: "No thought by that id" });
+      }
       const thoughtData = await Thought.updateOne(
         { _id: req.params.id },
         { $pull: { reactions: { reactionId: req.body.reactionId } } }
       );
-      if (!thoughtData) {
-        console.log("No thought by that id");
-        return res.status(500).json({ message: "No thought by that id" });
-      }
-      res.json(thoughtData);
+      res.json(thought);
     } catch (error) {
       console.log(error);
       res.status(500).json(error);
